Fix bug with unmodified sut arguments
When multiple sut arguments are used by a tested function,
and only some of them are modified, the old operator does
not change the others. For example in varray:

```
val append : 'a t -> 'a t -> 'a t
  (*@ t = append a b
      ensures t.contents = a.contents ++ b.contents *)
```

there is no point in wrapping `a.contents` or
`b.contents` with the `old` operator. However,
previously this resulted in `a` and `b` not being
rewritten with the particular SUT argument from the
stack of suts.

This error is not specific to returning suts, but
this behaviour was previously not exercised by any
of the test cases.